Refactor ContactSalesForm component to use constants for schema and options
