+++ src/components/ContactSalesForm.tsx
@@ -31,95 +31,14 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Loader2, CheckCircle, AlertCircle } from "lucide-react";
-
-// Disposable email domains for basic spam protection
-const disposableEmailDomains = [
-  "10minutemail.com",
-  "guerrillamail.com",
-  "mailinator.com",
-  "temp-mail.org",
-  "throwaway.email",
-];
-
-const contactSalesSchema = z.object({
-  companyName: z
-    .string()
-    .min(2, "Company name must be at least 2 characters")
-    .max(100, "Company name must be less than 100 characters"),
-  workEmail: z
-    .string()
-    .email("Please enter a valid email address")
-    .refine((email) => {
-      const domain = email.split("@")[1];
-      return !disposableEmailDomains.includes(domain?.toLowerCase());
-    }, "Please use a work email address"),
-  teamSize: z
-    .number({ required_error: "Team size is required" })
-    .int()
-    .min(1, "Team size must be at least 1")
-    .max(10000, "Team size must be less than 10,000"),
-  countryRegion: z.string().min(1, "Please select your country/region"),
-  hasConsent: z
-    .boolean()
-    .refine((val) => val === true, "You must agree to be contacted"),
-  phone: z
-    .string()
-    .optional()
-    .refine(
-      (val) => !val || val.length >= 10,
-      "Phone number must be at least 10 digits"
-    ),
-  useCase: z
-    .string()
-    .max(500, "Use case must be less than 500 characters")
-    .optional(),
-  preferredContactMethod: z.string().optional(),
-  preferredContactTime: z.string().optional(),
-  message: z
-    .string()
-    .max(1000, "Message must be less than 1000 characters")
-    .optional(),
-});
-
-type ContactSalesForm = z.infer<typeof contactSalesSchema>;
-
-const countries = [
-  "United States",
-  "Canada",
-  "United Kingdom",
-  "Germany",
-  "France",
-  "Australia",
-  "Japan",
-  "Singapore",
-  "Netherlands",
-  "Sweden",
-  "Norway",
-  "Denmark",
-  "Switzerland",
-  "Austria",
-  "Belgium",
-  "Finland",
-  "Ireland",
-  "New Zealand",
-  "South Korea",
-  "Other",
-];
-
-const contactMethods = ["Email", "Phone", "Video Call"];
-
-const timePreferences = [
-  "9 AM - 12 PM EST",
-  "12 PM - 3 PM EST",
-  "3 PM - 6 PM EST",
-  "9 AM - 12 PM PST",
-  "12 PM - 3 PM PST",
-  "3 PM - 6 PM PST",
-  "9 AM - 12 PM GMT",
-  "12 PM - 3 PM GMT",
-  "3 PM - 6 PM GMT",
-  "Flexible",
-];
+import {
+  ContactSalesForm,
+  contactSalesSchema,
+  countries,
+  contactMethods,
+  timePreferences,
+} from "@/constants/contact-enterprise";
+import { BACKEND_URL } from "@/auth/backend";
 
 interface ContactSalesDialogProps {
   children: React.ReactNode;
@@ -155,16 +74,13 @@ export function ContactSalesDialog({ children }: ContactSalesDialogProps) {
 
     try {
       // TODO: Replace with actual API endpoint URL from environment
-      const response = await fetch(
-        "http://localhost:3003/api/sales-contact/submit",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify(data),
-        }
-      );
+      const response = await fetch(`${BACKEND_URL}/sales-contact/submit`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(data),
+      });
 
       if (response.ok) {
         const result = await response.json();
